Fix silent float64 upcast in SummarizeCframe/SFrame/Rings/Spec output arrays
In the lvmdrp26 environment (numpy 1.26.4), np.nanpercentile always
computes and returns float64 regardless of the requested percentile, even
though the input FITS extensions are float32. -percent 50 dodged this by
using np.ma.median instead (a special-cased branch that does preserve
float32); any other -percent value fell through to nanpercentile and
silently doubled the size of every flux-like output extension.

Also fixed two related cases of the same bug class:
- SummarizeSframe.py's IVAR extension was float64 unconditionally (not
  just for -percent != 50), because of a "* 0.63694" Python-float
  multiply in the correction factor.
- SummarizeCframe.py/SummarizeSframe.py's new -by fiber mode has the same
  upcast via astropy.stats.sigma_clipped_stats (also always float64),
  affecting both the flux-like arrays and WAVE.

Rather than patch every percentile/sigma-clip call site, cast to float32
at the single point where each script stacks per-exposure results into
the final output array -- the same pattern SkySubSci.py/SummarizeSciSky.py
already used, which is why they were unaffected.

+++ py_progs/SummarizeCframe.py
@@ -468,11 +468,17 @@ def make_med_spec(xtab,data_dir,outfile='',percentile=50,exp_start=None,
         for col in meta_list[0]:
             xtab[col]=[m[col] for m in meta_list]
 
-    wav=np.array(wav)
-    xsci_flux=np.array(xsci_flux)
-    xsci_sky_e=np.array(xsci_sky_e)
-    xsci_sky_w=np.array(xsci_sky_w)
-    xsci_lsf=np.array(xsci_lsf)
+    # dtype=np.float32 here too: get_fiber_spec() upconverts wav to float64
+    # internally for mask-interpolation precision, but that shouldn't leak
+    # into the output -- pixel mode's wav is float32 already.
+    wav=np.array(wav,dtype=np.float32)
+    # dtype=np.float32 guards against np.nanpercentile/sigma_clipped_stats
+    # silently upcasting to float64 (numpy 1.26 does this for any
+    # percentile computation, even though the FITS extensions are float32)
+    xsci_flux=np.array(xsci_flux,dtype=np.float32)
+    xsci_sky_e=np.array(xsci_sky_e,dtype=np.float32)
+    xsci_sky_w=np.array(xsci_sky_w,dtype=np.float32)
+    xsci_lsf=np.array(xsci_lsf,dtype=np.float32)
     print(xsci_flux.shape,xsci_sky_e.shape,xsci_sky_w.shape)
     hdu1 = fits.PrimaryHDU(data=None)
     hdu1.header['Title'] = 'CFrame_Summmary'
